requirements.txt
torch==2.11.0
torchvision==0.26.0
transformers==5.6.2
faiss-cpu==1.13.2
numpy==2.2.6
pillow==12.2.0
streamlit==1.57.0
requests==2.33.1